Use uint64 instead of unsigned int to prevent overflows

+++ Vital.extension/Engine/public/texture.h
@@ -47,7 +47,7 @@ namespace Vital::Godot {
         
             struct Command {
                 Type type;
-                unsigned int tick;
+                uint64_t tick;
                 std::variant<Texture2D, SVG> payload;
             };
         protected:
